first fix, publish status, update version

+++ KingPim.Application/Repositories/Models/CategoryModel.cs
@@ -13,7 +13,7 @@ namespace KingPim.Application.Repositories.Models
         public DateTime DateCreated { get; set; }
         public DateTime DateUpdated { get; set; }
         public string EditedBy { get; set; }
-        public double Version { get; set; }
+        public int Version { get; set; }
         public bool PublishedStatus { get; set; }
     }
 }
