+++ KingPim.Application/Repositories/Models/ProductAttributeModel.cs
@@ -14,7 +14,7 @@ namespace KingPim.Application.Repositories.Models
         public DateTime DateCreated { get; set; }
         public DateTime DateUpdated { get; set; }
         public string EditedBy { get; set; }
-        public double Version { get; set; }
+        public int Version { get; set; }
         public bool PublishedStatus { get; set; }
         public int? SubCategoryId { get; set; }
         public virtual SubCategory SubCategory { get; set; }
