+++ KingPim.Application/Repositories/Models/SubCategoryModel.cs
@@ -13,7 +13,7 @@ namespace KingPim.Application.Repositories.Models
         public DateTime DateCreated { get; set; }
         public DateTime DateUpdated { get; set; }
         public string EditedBy { get; set; }
-        public double Version { get; set; }
+        public int Version { get; set; }
         public bool PublishedStatus { get; set; }
         public int? CategoryId { get; set; }
         public virtual Category Category { get; set; }
